requirements.txt
streamlit
pandas
Pillow
smtplib
ssl
email
python-dotenv
streamlit-authenticator
pyyaml
